chore(designdemo): replace old tab screens with unified bottom-sheet demo
- Replace bug.tsx, modals.tsx, keyboard-test.tsx tabs with a single
  bottom-sheet.tsx screen covering all openDrawer/editProp/openModal scenarios
- Update _layout.tsx to use new tab structure with icons
- Update package.json for expo-sdk upgrade (expo 54, RN 0.81.5, reanimated 4)
- Update metro.config.js: use blockList instead of deprecated blacklistRE,
  remove exclusionList dependency
