+++ examples/designdemo/package.json
@@ -5,7 +5,8 @@
   "scripts": {
     "start": "NODE_ENV=development expo start --dev-client --port 7765",
     "android": "NODE_ENV=development expo run:android --port 7765",
-    "ios": "NODE_ENV=development expo run:ios --port 7765  --device",
+    "ios": "NODE_ENV=development expo run:ios --port 7765 --device",
+    "start:device": "NODE_ENV=development expo start --dev-client --port 7765 --host lan",
     "web": "NODE_ENV=development expo start --web --port 7765",
     "open:android": "open -a \"Android Studio\" ./android",
     "open:ios": "open -a \"Xcode\" ./ios",
@@ -27,42 +28,45 @@
     "@formatjs/intl-numberformat": "^8.10.3",
     "@formatjs/intl-pluralrules": "^5.2.14",
     "@formatjs/intl-relativetimeformat": "^11.2.14",
-    "@gorhom/bottom-sheet": "^5.1.4",
+    "@gorhom/bottom-sheet": "^5.2.8",
     "@gorhom/portal": "^1.0.14",
-    "@react-native-async-storage/async-storage": "2.1.2",
-    "@react-native-community/slider": "4.5.6",
+    "@react-native-async-storage/async-storage": "2.2.0",
+    "@react-native-community/slider": "5.0.1",
     "@react-navigation/drawer": "^7.0.0",
     "@react-navigation/material-bottom-tabs": "^6.2.29",
     "@siteed/design-system": "workspace:^",
     "@siteed/react-native-logger": "workspace:^",
     "date-fns": "^3.6.0",
-    "expo": "^53.0.9",
-    "expo-constants": "~17.1.6",
-    "expo-linking": "~7.1.5",
-    "expo-localization": "~16.1.5",
-    "expo-router": "~5.0.7",
-    "expo-status-bar": "~2.2.3",
-    "expo-system-ui": "~5.0.7",
+    "expo": "^54",
+    "expo-constants": "~18.0.13",
+    "expo-dev-client": "~6.0.20",
+    "expo-linking": "~8.0.11",
+    "expo-localization": "~17.0.8",
+    "expo-router": "~6.0.23",
+    "expo-status-bar": "~3.0.9",
+    "expo-system-ui": "~6.0.9",
     "i18next": "^23.10.1",
     "intl-pluralrules": "^2.0.1",
-    "react": "19.0.0",
-    "react-dom": "19.0.0",
-    "react-native": "0.79.2",
-    "react-native-gesture-handler": "~2.24.0",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
+    "react-native": "0.81.5",
+    "react-native-edge-to-edge": "^1.7.0",
+    "react-native-gesture-handler": "~2.28.0",
     "react-native-localize": "^3.2.1",
     "react-native-paper": "^5.14.0",
     "react-native-paper-dates": "^0.22.25",
-    "react-native-reanimated": "~3.17.4",
-    "react-native-safe-area-context": "5.4.0",
-    "react-native-screens": "~4.10.0",
+    "react-native-reanimated": "~4.1.1",
+    "react-native-safe-area-context": "~5.6.0",
+    "react-native-screens": "~4.16.0",
     "react-native-vector-icons": "^10.0.3",
-    "react-native-web": "^0.20.0",
+    "react-native-web": "^0.21.0",
+    "react-native-worklets": "0.5.1",
     "testingui": "workspace:^"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
     "@types/gh-pages": "^6",
-    "@types/react": "~19.0.10",
+    "@types/react": "~19.1.10",
     "babel-plugin-module-resolver": "^5.0.2",
     "eslint": "^8.56.0",
     "eslint-config-universe": "^12.0.0",
@@ -71,7 +75,7 @@
     "rimraf": "^6.0.1",
     "serve": "^14.2.3",
     "ts-node": "^10.9.2",
-    "typescript": "^5.8.3"
+    "typescript": "~5.9.2"
   },
   "nohoist": [
     "**/@siteed/react-native-logger",
